añadida prueba complicada de la raiz cuadrada y refactorizando la función Raiz Cuadrada para usar las funciones propias en lugar de +,-,* y /

+++ src/rgb/Calculadora/interfaceCalculadora.java
@@ -41,9 +41,9 @@ public class interfaceCalculadora {
 	{
 		Calculadora calculadora = new Calculadora();
 		String respuesta = "S";
-		int val1;
-		int val2;
-		int resultadoInt = 0;
+		double val1;
+		double val2;
+		double resultadoInt = 0;
 		double resultatoDouble = 0;
 		BufferedReader lector = new BufferedReader(new InputStreamReader(System.in));
